// src/notification/config/comms.enum.ts
export enum CommsEventType {
  ORDER_CONFIRM = 'ORDER_CONFIRM',
  ORDER_SHIPPED = 'ORDER_SHIPPED',
  ORDER_DELIVERED = 'ORDER_DELIVERED',
  ORDER_DELIVERY_DELAYED = 'ORDER_DELIVERY_DELAYED',
  ORDER_CANCELLED = 'ORDER_CANCELLED',
  ORDER_FAILED = 'ORDER_FAILED',

  RETURN_INITIATED = 'RETURN_INITIATED',
  RETURN_CANCELLED = 'RETURN_CANCELLED',

  EXCHANGE_INITIATED = 'EXCHANGE_INITIATED',
  EXCHANGE_CANCELLED = 'EXCHANGE_CANCELLED',

  REFUND_INITIATED = 'REFUND_INITIATED',
}

export enum EmailMessageType {
  ORDER_CONFIRM = 'ORDER_CONFIRM',
  ORDER_SHIPPED = 'ORDER_SHIPPED',
  ORDER_DELIVERED = 'ORDER_DELIVERED',
  ORDER_DELIVERY_DELAYED = 'ORDER_DELIVERY_DELAYED',
  ORDER_CANCELLED = 'ORDER_CANCELLED',
  ORDER_FAILED = 'ORDER_FAILED',

  RETURN_INITIATED = 'RETURN_INITIATED',
  RETURN_CANCELLED = 'RETURN_CANCELLED',

  EXCHANGE_INITIATED = 'EXCHANGE_INITIATED',
  EXCHANGE_CANCELLED = 'EXCHANGE_CANCELLED',

  REFUND_INITIATED = 'REFUND_INITIATED',
}

export enum SmsMessageType {
  ORDER_CONFIRM = 'ORDER_CONFIRM',
  ORDER_CONFIRM_QTY_MORE = 'ORDER_CONFIRM_QTY_MORE',

  ORDER_SHIPPED_QTY_1 = 'ORDER_SHIPPED_QTY_1',
  ORDER_SHIPPED_QTY_2 = 'ORDER_SHIPPED_QTY_2',
  ORDER_SHIPPED_QTY_3_PLUS = 'ORDER_SHIPPED_QTY_3_PLUS',

  ORDER_DELIVERED_QTY_1 = 'ORDER_DELIVERED_QTY_1',
  ORDER_DELIVERED_QTY_2 = 'ORDER_DELIVERED_QTY_2',
  ORDER_DELIVERED_QTY_3_PLUS = 'ORDER_DELIVERED_QTY_3_PLUS',

  ORDER_DELIVERY_DELAYED_QTY_1 = 'ORDER_DELIVERY_DELAYED_QTY_1',
  ORDER_DELIVERY_DELAYED_QTY_2 = 'ORDER_DELIVERY_DELAYED_QTY_2',
  ORDER_DELIVERY_DELAYED_QTY_3_PLUS = 'ORDER_DELIVERY_DELAYED_QTY_3_PLUS',

  ORDER_CANCELLED = 'ORDER_CANCELLED',
  ORDER_FAILED = 'ORDER_FAILED',

  RETURN_INITIATED = 'RETURN_INITIATED',
  RETURN_CANCELLED = 'RETURN_CANCELLED',

  EXCHANGE_INITIATED = 'EXCHANGE_INITIATED',
  EXCHANGE_CANCELLED = 'EXCHANGE_CANCELLED',

  REFUND_INITIATED = 'REFUND_INITIATED',
}